+ Cast to unsigned int instead long for InvSqrt

+++ kex2/turok/math/mathlib.cpp
@@ -74,12 +74,12 @@ int kexMath::RoundPowerOfTwo(int x) {
 //
 
 float kexMath::InvSqrt(float x) {
-    long i;
+    unsigned int i;
     float r;
     float y;
     
     y = x * 0.5f;
-    i = *reinterpret_cast<long*>(&x);
+    i = *reinterpret_cast<unsigned int*>(&x);
     i = 0x5f3759df - (i >> 1);
     r = *reinterpret_cast<float*>(&i);
     r = r * (1.5f - r * r * y);
